Add thin panel aggregator
